Rename to reflect functionality.
